fix: fix disabled behavior and scriptLoader call
