+++ src/components/StripeCardInput/index.js
@@ -1,21 +1,124 @@
-import React, { useCallback } from 'react';
+import React, { useMemo, useEffect, useRef, useState } from 'react';
 import PropTypes from 'prop-types';
 import scriptLoader from 'react-async-script-loader';
-import CardInput from './component';
-
-const stripeApiV3Url = 'https://js.stripe.com/v3';
+import withReduxForm from '../../libs/hocs/withReduxForm';
+import { useTheme, useUniqueIdentifier, useLocale } from '../../libs/hooks';
+import RenderIf from '../RenderIf';
+import Label from '../Input/label';
+import HelpText from '../Input/styled/helpText';
+import ErrorText from '../Input/styled/errorText';
+import StyledContainer from './styled/container';
+import getError from './helpers/getErrror';
+import getCardElementOptions from './helpers/getCardElementOptions';
+import StyledCardInput from './styled/cardInput';
 
 /**
  * Stripe Card Input component are used for freeform data entry.
  * @category Form
  */
-const StripeCardInput = props => {
-    const { ...rest } = props;
+const StripeCardInput = React.forwardRef((props, ref) => {
+    const {
+        apiKey,
+        label,
+        hideLabel,
+        bottomHelpText,
+        error,
+        className,
+        style,
+        disabled,
+        required,
+        locale,
+        onChange,
+        onFocus,
+        onBlur,
+        tabIndex,
+        isScriptLoaded,
+        isScriptLoadSucceed,
+    } = props;
+    const [stripe, setStripe] = useState(null);
+    const cardRef = useRef();
+    const cardElementId = useUniqueIdentifier('card-element');
+    const theme = useTheme().rainbow;
+    const cardElementOptions = useMemo(() => getCardElementOptions(theme, disabled), [
+        disabled,
+        theme,
+    ]);
+    const localeStripe = useLocale(locale);
+    const elementOptions = useMemo(
+        () => ({
+            locale: localeStripe,
+            fonts: [
+                {
+                    cssSrc: 'https://fonts.googleapis.com/css?family=Lato&display=swap',
+                },
+            ],
+        }),
+        [localeStripe],
+    );
+    useEffect(() => {
+        if (isScriptLoaded && isScriptLoadSucceed && window.Stripe) {
+            setStripe(window.Stripe(apiKey));
+        }
+    }, [apiKey, isScriptLoadSucceed, isScriptLoaded]);
 
-    const Component = useCallback(scriptLoader(stripeApiV3Url)(CardInput), []);
+    // eslint-disable-next-line consistent-return
+    useEffect(() => {
+        if (stripe) {
+            const cardNode = cardRef.current;
+            const elements = stripe.elements(elementOptions);
+            const card = elements.create('card', cardElementOptions);
 
-    return <Component {...rest} />;
-};
+            card.mount(cardNode);
+            card.on('change', event => {
+                const stripeCardEvent = {
+                    stripe,
+                    card,
+                    isEmpty: event.empty,
+                    isComplete: event.complete,
+                    cardBrand: event.brand,
+                    error: getError(event.error),
+                };
+                onChange(stripeCardEvent);
+            });
+
+            card.on('focus', onFocus);
+            card.on('blur', onBlur);
+            const handlerBlur = () => card.blur();
+            const handlerFocus = () => card.focus();
+            cardNode.addEventListener('blur', handlerBlur);
+            cardNode.addEventListener('focus', handlerFocus);
+            return () => {
+                cardNode.removeEventListener('blur', handlerBlur);
+                cardNode.removeEventListener('focus', handlerFocus);
+                card.unmount();
+                card.destroy();
+            };
+        }
+    }, [cardElementOptions, elementOptions, onBlur, onChange, onFocus, stripe]);
+
+    return (
+        <StyledContainer disabled={disabled} className={className} style={style} ref={ref}>
+            <Label
+                label={label}
+                hideLabel={hideLabel}
+                inputId={cardElementId}
+                required={required}
+            />
+            <StyledCardInput
+                ref={cardRef}
+                id={cardElementId}
+                tabIndex={tabIndex}
+                disabled={disabled}
+            />
+            <RenderIf isTrue={!!bottomHelpText}>
+                <HelpText alignSelf="center">{bottomHelpText}</HelpText>
+            </RenderIf>
+            <RenderIf isTrue={!!error}>
+                <ErrorText alignSelf="center">{error}</ErrorText>
+            </RenderIf>
+        </StyledContainer>
+    );
+});
 
 StripeCardInput.propTypes = {
     /** The application's API key. To use Stripe,
@@ -89,4 +192,6 @@ StripeCardInput.defaultProps = {
     style: undefined,
 };
 
-export default StripeCardInput;
+export default scriptLoader('https://js.stripe.com/v3')(withReduxForm(StripeCardInput));
+
+export { StripeCardInput as Component };
