+++ src/components/StripeCardInput/styled/cardInput.js
@@ -7,6 +7,7 @@ const StyledCardInput = attachThemeAttrs(styled.div)`
     background-color: ${props => props.palette.background.main};
     border: 1px solid ${props => props.palette.border.main};
     border-radius: ${BORDER_RADIUS_2};
+    cursor: text;
     ${props =>
         props.disabled &&
         `
@@ -14,6 +15,13 @@ const StyledCardInput = attachThemeAttrs(styled.div)`
         border: 1px solid ${props.palette.border.disabled} !important;
         cursor: not-allowed;
         user-select: none;
+
+        &:focus,
+            &:active {
+                box-shadow: none;
+                background-color: ${props.palette.background.disabled} !important;
+                border: 1px solid ${props.palette.border.disabled} !important;
+            }
     `}
 `;
